--- components/molecules/calendarDatePicker/README.md
@@ -1,3 +1,0 @@
-# Calendar Date Picker
-
-Provides select to pick the day month and year
