+++ components/molecules/index.js
@@ -17,6 +17,7 @@ export * from './dropzone'
 export * from './emojiMart'
 export * from './getAddress'
 export * from './hero'
+export * from './inputPicker'
 export * from './imageLocation'
 export * from './intercom'
 export * from './location'
