+++ components/molecules/inputPicker/README.md
@@ -1,0 +1,3 @@
+# Date Picker
+
+Provides three selects to pick the day month and year
